Suggestion: Removing the border around the image
The border is a bit distracting and thus annoying (for me at least) when using images with transparency on the sides. This would remove the border. In most cases this should only be noticeable on images with transparency on the sides. On images without transparency it should hardly be noticeable.